ComponentTreeRoot is now in its own file
